Deduplicate timer wakeup impls
