+++ esp-hal/src/rtc_cntl/sleep/esp32s2.rs
@@ -1,7 +1,6 @@
 use super::{
     Ext0WakeupSource,
     Ext1WakeupSource,
-    TimerWakeupSource,
     UlpWakeupSource,
     WakeSource,
     WakeTriggers,
@@ -91,37 +90,6 @@ impl WakeSource for UlpWakeupSource {
         // This one needs to be false to keep the ULP timer and ULP GPIO happy!
         // Possibly relevant issue: https://github.com/espressif/esp-idf/issues/10595
         sleep_config.set_rtc_peri_pd_en(false);
-    }
-}
-
-impl WakeSource for TimerWakeupSource {
-    fn apply(
-        &self,
-        rtc: &Rtc<'_>,
-        triggers: &mut WakeTriggers,
-        _sleep_config: &mut RtcSleepConfig,
-    ) {
-        triggers.set_timer(true);
-        let rtc_cntl = LPWR::regs();
-        // TODO: maybe add check to prevent overflow?
-        let ticks = crate::clock::us_to_rtc_ticks(self.duration.as_micros());
-        // "alarm" time in slow rtc ticks
-        let now = rtc.time_since_boot_raw();
-        let time_in_ticks = now + ticks;
-        unsafe {
-            rtc_cntl
-                .slp_timer0()
-                .write(|w| w.slp_val_lo().bits((time_in_ticks & 0xffffffff) as u32));
-
-            rtc_cntl
-                .int_clr()
-                .write(|w| w.main_timer().clear_bit_by_one());
-
-            rtc_cntl.slp_timer1().write(|w| {
-                w.slp_val_hi().bits(((time_in_ticks >> 32) & 0xffff) as u16);
-                w.main_timer_alarm_en().set_bit()
-            });
-        }
     }
 }
